tests: sync hardware-loop perf benchmark with raw lp.setup

+++ tests/gem5/riscv_ai_ext/hwloop_perf_src/hwloop_redirect_ref.cpp
@@ -24,7 +24,7 @@ runRefKernel(perf::hwloop::KernelState &state)
 extern "C" [[noreturn]] void
 _start()
 {
-    const std::uint64_t checksum = perf::hwloop::runBenchmark(runRefKernel);
+    const std::uint32_t checksum = perf::hwloop::runBenchmark(runRefKernel);
     perf::hwloop::printBenchmarkChecksum(checksum);
     perf::syscallExit(0);
 }
